Creates git sample code for Apps Script examples.
go/36579432 - b/36579432

This CL follows the plan in the shared design doc.
I've consulted with Cloud DPEs and googlesamples repo maintainers w.r.t. this design.

None of the gs code has been modified besides the comments, and structure of the code.
Previously, our Apps Script sample snippets were hosted in 4 formats:

- Plaintext within markdown
- Files on google3
- Gerrit
- GitHub

This is an effort in enabling the current samples to work well with the CLI.

The CLI will work with:
- Bound and non-bound scripts
- Scripts small and large
- Projects in subdirectories of git repos

Please see details in the design doc for high-level discussion.

Change-Id: Ic2f8bca8a32b9894e0a5b1e227e0545b7c195877