Update adsense.gs for adsense v2.
This updates the sample code to use v2 of the AdSense Management API
rather than deprecated-and-soon-to-be-removed v1.4, including adding a
new sample to list accounts (the v1.4 API picked an account if one was
not specified; v2 requires the account to be specified explicitly).

This also modernizes some of the JavaScript used here, and formats the
code using semistandard.