Update removingDuplicates.gs
Fix magic behavior in sometimes.
I spent about 2 hours to understand why when I run this function, it gave me magic. 
And when I tried to debug it in console.log I also got magic. 
Also you can understand that using 'for in ...' for arrays is a bad practice - https://stackoverflow.com/questions/500504/why-is-using-for-in-for-array-iteration-a-bad-idea